--- src/main/java/domain/Category.java
@@ -1,17 +1,0 @@
-package domain;
-
-public enum Category {
-    CHICKEN("치킨"),
-    BEVERAGE("음료");
-
-    private final String name;
-
-    Category(final String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return "[" + name + "]";
-    }
-}
