+++ src/main/java/domain/table/TableRepository.java
@@ -1,8 +1,10 @@
-package domain;
+package domain.table;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+
+import domain.table.order.Order;
 
 public class TableRepository {
 	private static final List<Table> tables = new ArrayList<>();
