+++ src/test/java/domain/TablesTest.java
@@ -5,6 +5,13 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import domain.table.Table;
+import domain.table.TableRepository;
+import domain.table.Tables;
+import domain.table.order.OrderCount;
+import domain.table.order.menu.Category;
+import domain.table.order.menu.Menu;
+
 class TablesTest {
 
 	@DisplayName("테이블번호에 맞는 테이블을 가져온다.")
@@ -12,7 +19,7 @@ class TablesTest {
 	void getOrThrow() {
 		Tables tables = new Tables(TableRepository.tables());
 
-		Table table = tables.getTable(1);
+		Table table = tables.findTable(1);
 
 		assertThat(table).isNotNull();
 	}
@@ -22,7 +29,7 @@ class TablesTest {
 	void name() {
 		Tables tables = new Tables(TableRepository.tables());
 
-		assertThatThrownBy(() -> tables.getTable(99))
+		assertThatThrownBy(() -> tables.findTable(99))
 			.isInstanceOf(IllegalArgumentException.class)
 			.hasMessageContaining("찾을 수 없는 테이블입니다.");
 	}
@@ -32,13 +39,13 @@ class TablesTest {
 	void getOrderedTable() {
 		//given
 		Tables tables = new Tables(TableRepository.tables());
-		Table table = tables.getTable(1);
+		Table table = tables.findTable(1);
 
 		Menu menu = new Menu(1, "치킨", Category.CHICKEN, 2_000);
 		table.addOrder(menu, OrderCount.of(5));
 
 		//when
-		Table actual = tables.getOrderedTable(1);
+		Table actual = tables.findOrderedTable(1);
 
 		assertThat(actual).isNotNull();
 	}
@@ -48,7 +55,7 @@ class TablesTest {
 	void name3() {
 		Tables tables = new Tables(TableRepository.tables());
 
-		assertThatThrownBy(() -> tables.getOrderedTable(3))
+		assertThatThrownBy(() -> tables.findOrderedTable(3))
 			.isInstanceOf(IllegalArgumentException.class)
 			.hasMessageContaining("주문되지 않는 테이블입니다.");
 	}
